+++ packages/devtools/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-alpha.24](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@8.0.0-alpha.23...@react-navigation/devtools@8.0.0-alpha.24) (2026-06-11)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+# [8.0.0-alpha.23](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@8.0.0-alpha.22...@react-navigation/devtools@8.0.0-alpha.23) (2026-06-10)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
 # [8.0.0-alpha.22](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@8.0.0-alpha.21...@react-navigation/devtools@8.0.0-alpha.22) (2026-06-09)
 
 **Note:** Version bump only for package @react-navigation/devtools